+++ src/main/java/com/atlauncher/thread/PasteUpload.java
@@ -29,6 +29,7 @@ import java.net.URLEncoder;
 import java.util.concurrent.Callable;
 
 import com.atlauncher.App;
+import com.atlauncher.Network;
 import com.atlauncher.constants.Constants;
 import com.atlauncher.managers.LogManager;
 
@@ -58,6 +59,7 @@ public final class PasteUpload implements Callable<String> {
         }
         conn.setDoOutput(true);
         try {
+            conn.addRequestProperty("User-Agent", Network.USER_AGENT);
             conn.connect();
             conn.getOutputStream().write(urlParameters.getBytes());
             conn.getOutputStream().flush();