+++ CHANGELOG.md
@@ -3,12 +3,10 @@
 This changelog only contains the changes that are unreleased. For changes for individual releases, please visit the
 [releases](https://github.com/ATLauncher/ATLauncher/releases) page on GitHub.
 
-## 3.4.6.4
+## 3.4.6.5
 
 ### New Features
 
 ### Fixes
-- Issue with checking mods on CurseForge causing a NPE
-- System out interceptor not working for errors or showing info
 
 ### Misc